Update controller from Nguyen Le to Hieu branch

+++ kabina_project_api/src/main/java/Kabina/Service/UserService.java
@@ -8,7 +8,7 @@ public interface UserService {
 	Iterable<User> findAll();
 	public User addUser(User usr);
 	public User updateUser(User usr);
-	public void deleteUser (Integer Id);
+	public void deleteUser (Long Id);
 	int checkUserExist(String userName);
 	Map<String, String> getUserProfile(long userId);
 	User findByUserName(String userName);
